squash another dtype bug

+++ agents/utils.py
@@ -10,11 +10,11 @@ class ReplayBuffer(object):
     self.ptr = 0
     self.size = 0
 
-    self.state = np.zeros((max_size, state_dim))
-    self.action = np.zeros((max_size, action_dim))
-    self.next_state = np.zeros((max_size, state_dim))
-    self.reward = np.zeros((max_size, 1))
-    self.not_done = np.zeros((max_size, 1))
+    self.state = np.zeros((max_size, state_dim), dtype=np.float32)
+    self.action = np.zeros((max_size, action_dim), dtype=np.float32)
+    self.next_state = np.zeros((max_size, state_dim), dtype=np.float32)
+    self.reward = np.zeros((max_size, 1), dtype=np.float32)
+    self.not_done = np.zeros((max_size, 1), dtype=np.float32)
 
     self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
